// src/app/utils/enums/advanceSearchKeyEnum.ts
export enum AdvanceSearchKeyEnum {
    Branch = 'branch',
    CountryCode = 'countryCode',
    CompanyNameInter = 'companyNameInter',
    CompanyNameLocal = 'companyNameLocal',
    ProductCategoryLev1 = 'productCategoryLev1',
    ProductCategoryLev2 = 'productCategoryLev2',
    ProductCategoryLev3 = 'productCategoryLev3',
    ProductCategoryAVL_Lev1 = 'productCategoryAVL_Lev1',
    ProductCategoryAVL_Lev2 = 'productCategoryAVL_Lev2',
    ProductCategoryAVL_Lev3 = 'productCategoryAVL_Lev3',
    GeneralName = 'generalName',
    RegisteredCapital = 'registeredCapital',
    YearEstablished = 'yearEstablished',
    BusinessEntity = 'businessEntity',
    BusinessType = 'businessType',
    ProductKeyword = 'keyword',
    InvNameInter = 'invNameInter',
    InvNameLocal = 'invNameLocal',
    OrgId = 'orgId',
    IsPTVNVerified = 'isPtvnVerified',
    SdFavoriteEid = 'sdFavoriteEid',
    SdFavoritesSysUserId = 'sdFavoritesSysUserId',
    SdFavoriteTenantId = 'sdFavoriteTenantId',
    POOnline = 'poOnline',
    IsDelete = 'isDelete',
    Other = 'other',
    QualifiedSupplierStatus = 'qualifiedSupplierStatus',
    IsPassCreateAVL = 'isPassCreateAVL',
}